move relevant files to craftassist_cuberite_utils
