+++ droidlet/interpreter/craftassist/schematic_helper.py
@@ -8,7 +8,7 @@ import numpy as np
 from typing import List, Tuple, Union, Optional
 
 # TODO with subinterpret
-from droidlet.lowlevel.minecraft import block_data
+from droidlet.lowlevel.minecraft.craftassist_cuberite_utils import block_data
 from droidlet.interpreter.craftassist import size_words
 from .block_helpers import get_block_type
 from droidlet.base_util import number_from_span
