+++ droidlet/lowlevel/minecraft/craftassist_cuberite_utils/ping_cuberite.py
@@ -12,7 +12,7 @@ PONG = b"\x09\x01\x00\x00\x00\x00\x00\x00\x00\x2a"
 
 
 def ping(host, port, timeout=10):
-    """Ping craftassist_cuberite using a socket connection"""
+    """Ping craftassist_cuberite_utils using a socket connection"""
     s = socket.socket()
     s.connect((host, port))
     s.settimeout(timeout)
