+++ droidlet/lowlevel/minecraft/craftassist_cuberite_utils/wait_for_cuberite.py
@@ -1,10 +1,8 @@
 """
 Copyright (c) Facebook, Inc. and its affiliates.
 """
-
 import socket
 import time
-
 import ping_cuberite
