+++ droidlet/perception/craftassist/heuristic_perception.py
@@ -17,7 +17,7 @@ from droidlet.lowlevel.minecraft.mc_util import (
     to_block_pos,
     fill_idmeta,
 )
-from droidlet.lowlevel.minecraft.block_data import BORING_BLOCKS, PASSABLE_BLOCKS, COLOR_BID_MAP
+from droidlet.lowlevel.minecraft.craftassist_cuberite_utils.block_data import BORING_BLOCKS, PASSABLE_BLOCKS, COLOR_BID_MAP
 from droidlet.perception.craftassist.search import depth_first_search
 from droidlet.memory.craftassist.mc_memory_nodes import InstSegNode, BlockObjectNode
